6_metrics/mlservice/pyproject.toml
[project]
name = "mlapp"
version = "0.1.0"
description = "mlops homework app"
authors = []

requires-python = ">=3.11,<3.13"

dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    "uvloop>=0.19,<1",
    "mlflow>=3.8,<4",
    "numpy>=1.26,<3",
    "scikit-learn>=1.4,<2",
    "psycopg2-binary>=2.9.9,<3",
    "starlette_exporter==0.23.0",
]